Basic chart operational (Buy price only)
There is still an issue with the loading process since it can cause confusion for the user when old data is being presented for a different selection.  A state system will have to be used taking advantage of the TargetCrypto field ID since that's probably the only way to differentiate the datasets.